Extract useVideoKeyboard and useVideoMetadata hooks from Home
